fix(mcp-server): SMI-5949 adversarial security review fixes
Fixes every finding from the Opus adversarial review (verdict:
SHIP-WITH-FIXES). Self-approval enforcement, RLS predicate completeness,
credential-tier correctness, and both prior implementation discoveries
were confirmed clean under adversarial tracing -- these fixes are all at
the edges: stub parity, message accuracy, audit trail, changelog.

H-1 (High): registry-tools.stub.ts's getContent() was missing the
approvalStatus==='approved' filter that list()/get() already had --
the cross-transport test was asserting "a pending, unapproved version
installs successfully" as a passing invariant, the exact inverse of the
feature. Fixed the stub, approved fixtures across cross-transport.test.ts
and install-action.test.ts (which had the identical gap, found while
fixing the first), and added a genuinely non-vacuous negative test --
empirically confirmed it fails with the fix mentally reverted.

M-1: not a bug -- traced to Wave 1's own migration comment
(20260809000000:78-86), which already documents and justifies "an admin
can only deprecate an APPROVED row" as intentional RLS behavior. Fixed
only the stale application-layer message/doc-comment that didn't reflect
this already-shipped, already-decided design.

M-2: deprecate/undeprecate not-found errors now use the shared
registrySkillNotFoundMessage() helper instead of a hand-written
duplicate.

M-3: corrected a false "admin only" claim about includeDeprecated --
list() runs on the service-role/license-key path with no auth.uid(), so
any team member holding the license key can pass it, not just admins.
Fixed in both the code message and the plan doc's smoke table.

M-4: registry-tools.cross-transport.test.ts's stub/live parity fixtures
for notAdmin/terminal were hand-written paraphrases that didn't match
what review_private_registry_submission actually raises. Replaced with
the migration's verbatim text (read directly from the .sql file), split
into a new registry-tools.review-parity.test.ts to stay under the
500-line gate.

M-5: publish() now audits result:'error' with detail:'readback_failed'
before rethrowing when the post-insert read-back RPC fails -- previously
a committed row with a failed read-back left zero audit trail.

M-6: CHANGELOG.md gains the missing Wave 2 [Unreleased] entry (the
actual breaking change -- publish now requires skillsmith login,
unconditional from merge day); corrected a now-false "unshipped
(SMI-5949)" line that contradicted the new entry a few lines away.

O-1 (Low, docs-only): documented the immutability error's narrow,
same-team-only existence-oracle as an accepted limitation inherent to
the already-shipped UNIQUE constraint -- no behavior change.

L-1 (Low): filed as SMI-5989 (get_private_registry_submissions needs a
deprecated column) -- an RPC RETURNS TABLE widening is a real migration,
out of scope for this app-layer PR. submissions()'s deprecated:false is
documented as a known limitation pending that follow-up.

L-2: mapSubmissionRow() now nulls registryUrl for non-approved rows,
matching what the pending-branch message text already correctly implied.

nit: bindUserClient's generic "Failed to X skill" message now reads
correctly for submissions/approve/reject too.

No .sql migration touched (confirmed via git diff --name-only -- '*.sql'
returning empty) -- every fix here is app-layer, test, or docs. Full
mcp-server suite: 2854 passing, zero regressions (independently
re-verified by the coordinating session, not just trusted from the
agent report).

Refs SMI-5949, SMI-5989
